+++ mux/src/look.cpp
@@ -1,6 +1,6 @@
 // look.cpp -- Commands which look at things.
 //
-// $Id: look.cpp,v 1.34 2003-01-21 01:33:51 sdennis Exp $
+// $Id: look.cpp,v 1.35 2003-01-31 06:34:20 sdennis Exp $
 //
 // MUX 2.2
 // Copyright (C) 1998 through 2003 Solid Vertical Domains, Ltd. All
@@ -2382,7 +2382,7 @@ void do_decomp
     // Strip out ANSI in one place rather than have it done in
     // several places.
     //
-    unsigned int len;
+    size_t len;
     char *p = strip_ansi(thingname, &len);
     memcpy(thingname, p, len+1);
