+++ mux/src/stringutil.cpp
@@ -1,6 +1,6 @@
 // stringutil.cpp -- string utilities.
 //
-// $Id: stringutil.cpp,v 1.35 2003-01-14 17:55:54 sdennis Exp $
+// $Id: stringutil.cpp,v 1.36 2003-01-31 06:34:20 sdennis Exp $
 //
 // MUX 2.2
 // Copyright (C) 1998 through 2003 Solid Vertical Domains, Ltd. All
@@ -499,7 +499,7 @@ int ANSI_lex(int nString, const char *pString, int *nLengthToken0, int *nLengthT
     }
 }
 
-char *strip_ansi(const char *szString, unsigned int *pnString)
+char *strip_ansi(const char *szString, size_t *pnString)
 {
     static char Buffer[LBUF_SIZE];
     char *pBuffer = Buffer;
@@ -1793,7 +1793,7 @@ BOOL minmatch(char *str, char *target, int min)
 // --------------------------------------------------------------------------
 // StringCloneLen: allocate memory and copy string
 //
-char *StringCloneLen(const char *str, unsigned int nStr)
+char *StringCloneLen(const char *str, size_t nStr)
 {
     char *buff = (char *)MEMALLOC(nStr+1);
     if (ISOUTOFMEMORY(buff))
